moved plugins to correct location
